+++ hdl/avr/mega/2560/include/reg.hpp
@@ -1,0 +1,6 @@
+#ifndef HDL_AVR_MEGA_2560_REG_HPP
+#define HDL_AVR_MEGA_2560_REG_HPP
+
+#include "../../xx0_1/include/reg.hpp"
+
+#endif//HDL_AVR_MEGA_2560_REG_HPP
