+++ hdl/common/include/hdl/reg.hpp
@@ -1,0 +1,9 @@
+#ifndef HDL_COMMON_REG_HPP
+#define HDL_COMMON_REG_HPP
+
+
+#if defined __AVR_ATmega2560__
+#include "../../../avr/mega/2560/include/reg.hpp"
+#endif    //__AVR_ATmega2560__
+
+#endif//HDL_COMMON_REG_HPP
