+++ hdl/common/include/util/memory_mapping.hpp
@@ -1,7 +1,7 @@
 #ifndef MEMORY_MAPPING_HPP
 #define MEMORY_MAPPING_HPP
 
-#include "integer.hpp"
+#include <base/integer.hpp>
 
 namespace MemoryMapping {
 
@@ -49,18 +49,18 @@ namespace MemoryMapping {
 
 }
 
-#include "register.hpp"
-#include "config.hpp"
-#include "field.hpp"
-#include "bitfield.hpp"
-#include "virtual_field.hpp"
-#include "virtual_config.hpp"
+#include "memory_mapping/register.hpp"
+#include "memory_mapping/config.hpp"
+#include "memory_mapping/field.hpp"
+#include "memory_mapping/bitfield.hpp"
+#include "memory_mapping/virtual_field.hpp"
+#include "memory_mapping/virtual_config.hpp"
 
-#include "register_impl.hpp"
-#include "config_impl.hpp"
-#include "field_impl.hpp"
-#include "bitfield_impl.hpp"
-#include "virtual_field_impl.hpp"
-#include "virtual_config_impl.hpp"
+#include "memory_mapping/register_impl.hpp"
+#include "memory_mapping/config_impl.hpp"
+#include "memory_mapping/field_impl.hpp"
+#include "memory_mapping/bitfield_impl.hpp"
+#include "memory_mapping/virtual_field_impl.hpp"
+#include "memory_mapping/virtual_config_impl.hpp"
 
 #endif//MEMORY_MAPPING_HPP
