+++ hdl/common/include/util/memory_mapping/bitfield.hpp
@@ -1,8 +1,6 @@
 #ifndef BITFIELD_HPP
 #define BITFIELD_HPP
 
-#include "memory_mapping.hpp"
-
 namespace MemoryMapping {
 
   template<typename RegType, typename Group, int BITNUM>
