+++ hdl/common/include/util/memory_mapping/field.hpp
@@ -1,8 +1,6 @@
 #ifndef FIELD_HPP
 #define FIELD_HPP
 
-#include "memory_mapping.hpp"
-
 namespace MemoryMapping {
 
   template<typename RegType, typename Group, RegType MASK>
