+++ hdl/common/include/util/memory_mapping/field_impl.hpp
@@ -1,8 +1,6 @@
 #ifndef FIELD_IMPL_HPP
 #define FIELD_IMPL_HPP
 
-#include "memory_mapping.hpp"
-
 namespace MemoryMapping {
 
   template<typename RegType, typename Group, RegType MASK>
