+++ hdl/common/include/util/memory_mapping/register_decl_arith.hpp
