+++ hdl/common/include/util/memory_mapping/register_impl.hpp
@@ -1,8 +1,6 @@
 #ifndef REGISTER_IMPL_HPP
 #define REGISTER_IMPL_HPP
 
-#include "memory_mapping.hpp"
-
 namespace MemoryMapping {
 
   template<typename RegType, typename Group>
