+++ hdl/common/include/util/memory_mapping/virtual_config.hpp
@@ -1,8 +1,7 @@
 #ifndef VIRTUAL_CONFIG_HPP
 #define VIRTUAL_CONFIG_HPP
 
-#include "memory_mapping.hpp"
-#include "static_list.hpp"
+#include <base/static_list.hpp>
 
 namespace MemoryMapping {
