+++ hdl/common/include/util/memory_mapping/virtual_field_impl.hpp
@@ -1,8 +1,6 @@
 #ifndef VIRTUAL_FIELD_IMPL_HPP
 #define VIRTUAL_FIELD_IMPL_HPP
 
-#include "memory_mapping.hpp"
-
 namespace MemoryMapping {
 
   template<typename Field, typename ... Next>
